Revert refactor changes
